examples/ava/requirements.txt
ffmpeg-python
h5py
numpy
opencv-python
PyYAML
